style: adiciona estilizações para pagina de favoritos e de receitas feitas
