+++ .stylelintrc.json
@@ -69,7 +69,6 @@
     "linebreaks": "unix",
     "max-empty-lines": 2,
     "max-line-length": 90,
-    "max-nesting-depth": 2,
     "media-feature-colon-space-after": "always",
     "media-feature-colon-space-before": "always",
     "media-feature-name-case": "lower",
@@ -88,7 +87,6 @@
     "no-extra-semicolons": true,
     "no-invalid-double-slash-comments": true,
     "no-missing-end-of-source-newline": true,
-    "no-unknown-animations": true,
     "number-leading-zero": "always",
     "number-max-precision": 3,
     "number-no-trailing-zeros": true,
@@ -111,11 +109,9 @@
     "selector-combinator-space-after": "always",
     "selector-combinator-space-before": "always",
     "selector-descendant-combinator-no-non-space": true,
-    "selector-list-comma-space-after": "always",
     "selector-list-comma-space-before": "never",
     "selector-max-attribute": 3,
     "selector-max-class": 5,
-    "selector-max-combinators": 3,
     "selector-max-compound-selectors": 5,
     "selector-max-empty-lines": 0,
     "selector-max-id": 5,
