+++ src/components/Footer/_footer.scss
@@ -7,7 +7,6 @@
   display: flex;
   height: 55px;
   justify-content: space-around;
-  max-width: 1366px;
   position: fixed;
   width: 100%;
   z-index: 1;
