+++ src/css/styles.scss
@@ -4,7 +4,7 @@
 
 @import '../components/global';
 
-@import '../pages/styles/global';
+@import '../pages/global';
 
 * {
   box-sizing: border-box;
