+++ src/pages/_global.scss
@@ -1,0 +1,3 @@
+@import './styles/global';
+
+@import './myRecipes/global';
