+++ src/pages/myRecipes/FavoritesRecipes/FavoriteRecipesCard/_favoriteRecipesCard.scss
@@ -1,0 +1,56 @@
+.container-favorites {
+
+  .cards-container {
+    align-items: center;
+    display: flex;
+    flex-wrap: wrap;
+    max-width: 1142px;
+    padding: 0 1rem;
+    width: 100%;
+
+    .container-card {
+      background-color: $white;
+      border-radius: 0.5rem;
+      box-shadow: 2px 2px 5px gray;
+      display: flex;
+      height: 150px;
+      margin: 10px;
+      min-width: 350px;
+      padding-right: 0.25rem;
+      position: relative;
+      width: max-content;
+
+      .container-text-n-icons {
+        display: flex;
+        flex-direction: column;
+        gap: 0.5rem;
+        padding-top: 0.5rem;
+        position: relative;
+        width: 50%;
+
+        p {
+          color: gray;
+          font-size: 0.85rem;
+        }
+
+        h3 {
+          font-size: 1.4rem;
+        }
+      }
+
+      a {
+        color: $green-leaf;
+        font-weight: 700;
+        margin: 0;
+        padding: 0;
+
+        img {
+          border-radius: 0.5rem;
+          height: 150px;
+          padding-right: 15px;
+          width: 100%;
+        }
+      }
+    }
+  }
+}
