+++ src/pages/myRecipes/_global.scss
@@ -1,0 +1,5 @@
+@import './FavoritesRecipes/FavoriteRecipesCard/favoriteRecipesCard';
+
+@import './MadeRecipes/MadeRecipesCard/madeRecipesCard';
+
+@import './HandleFavoriteAndShare/handleFavoriteAndShare';
