+++ src/pages/styles/_detailsRecipe.scss
@@ -33,6 +33,8 @@
     display: flex;
     flex-direction: column;
     justify-content: space-between;
+    margin: 0 auto;
+    max-width: 1142px;
 
     h1 {
       color: $green-leaf;
@@ -46,6 +48,8 @@
   }
 
   .container-details-ingredients {
+    margin: 0 auto;
+    max-width: 1142px;
 
     .recipe-ingredients {
       padding: 20px 0;
@@ -60,6 +64,8 @@
   }
 
   .instructions {
+    margin: 0 auto;
+    max-width: 1142px;
     padding: 10px;
     text-align: justify;
     width: 100%;
@@ -78,6 +84,8 @@
   .recommendations-recipes {
     display: flex;
     gap: 3px;
+    margin: 0 auto;
+    max-width: 1142px;
     overflow-x: scroll;
     width: 100%;
