+++ src/pages/styles/_global.scss
@@ -3,3 +3,5 @@
 @import './detailsRecipe';
 
 @import 'progressRecipe';
+
+@import './myRecipes';
