+++ src/pages/styles/_myRecipes.scss
@@ -1,0 +1,29 @@
+.container-favorites,
+.container-made {
+  align-items: center;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  width: 100%;
+
+  .filter-container {
+    align-items: center;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 0.25rem;
+    height: 60px;
+    justify-content: space-evenly;
+    margin: 0 auto;
+    max-width: 375px;
+    padding: 1rem 0;
+    width: 100%;
+
+    button {
+
+      @include generic-button;
+      color: white;
+      padding: 0.25rem 0.5rem;
+      width: 30%;
+    }
+  }
+}
