+++ src/pages/styles/_recipes.scss
@@ -24,6 +24,8 @@
   flex-wrap: wrap;
   gap: 1rem;
   justify-content: center;
+  margin: 0 auto;
+  max-width: 1160px;
   overflow-y: auto;
   padding-bottom: 70px;
   width: 100%;
