--- src/pages/styles/favoriteRecipes.css
@@ -1,21 +1,0 @@
-.container-favorites {
-  min-height: 100vh;
-  width: 100vw;
-}
-
-.filter-favorite-btn {
-  background-color: #673;
-  border: none;
-  border-radius: 5px;
-  color: white;
-  margin: 10px 5px;
-  padding: 3px 0;
-  width: 30%;
-}
-
-.container-filters {
-  display: flex;
-  justify-content: center;
-  text-align: center;
-  width: 100%;
-}
